LPD-47689 Move since this does not affect DB Partition only
